fix(pdf): Arregla errores varios (#27)

+++ src/main/java/mathtexpedia/es/api/presentation/pdf/PDFController.java
@@ -1,13 +1,16 @@
 package mathtexpedia.es.api.presentation.pdf;
 
 import mathtexpedia.es.api.domain.exception.MathtexpediaInvalidException;
+import mathtexpedia.es.api.domain.exception.MathtexpediaUnauthorizedException;
 import mathtexpedia.es.api.domain.model.pdf.CreatePDFDto;
+import mathtexpedia.es.api.domain.security.UserProfile;
 import mathtexpedia.es.api.persistence.pdf.PDF;
 import mathtexpedia.es.api.presentation.GenericController;
 import mathtexpedia.es.api.service.pdf.PDFService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -30,8 +33,11 @@ public class PDFController extends GenericController {
     }
 
     @PostMapping("/create")
-    public ResponseEntity<PDF> createPDF(@RequestBody CreatePDFDto pdf) throws MathtexpediaInvalidException {
+    public ResponseEntity<PDF> createPDF(@RequestBody CreatePDFDto pdf,
+                                         @AuthenticationPrincipal UserProfile user) throws MathtexpediaInvalidException, MathtexpediaUnauthorizedException {
         logger.debug("Called recieve to create PDF {}",pdf);
+
+        checkIfAdmin(user);
 
         PDF created = pdfService.createPDF(pdf);
         return ResponseEntity.status(HttpStatus.CREATED).body(created);
@@ -39,16 +45,22 @@ public class PDFController extends GenericController {
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<PDF> deletePDF(@RequestParam String pdfName) throws MathtexpediaInvalidException {
+    public ResponseEntity<PDF> deletePDF(@RequestParam String pdfName,
+                                         @AuthenticationPrincipal UserProfile user) throws MathtexpediaInvalidException, MathtexpediaUnauthorizedException {
         logger.debug("Called recieve to delete PDF {}",pdfName);
+
+        checkIfAdmin(user);
 
         pdfService.deletePDF(pdfName);
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(null);
     }
 
     @PatchMapping("/update")
-    public ResponseEntity<PDF> updatePDF(@RequestBody PDF pdf, @RequestParam long pdfId) throws MathtexpediaInvalidException {
+    public ResponseEntity<PDF> updatePDF(@RequestBody PDF pdf, @RequestParam long pdfId,
+                                         @AuthenticationPrincipal UserProfile user) throws MathtexpediaInvalidException, MathtexpediaUnauthorizedException {
         logger.debug("Called recieve to update PDF {}",pdf);
+
+        checkIfAdmin(user);
 
         pdf = pdfService.updatePDF(pdf, pdfId);
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(pdf);
